added a partial implementation of control
