+++ main.go
@@ -5,25 +5,29 @@ import (
 	"fmt"
 	"net/http"
 
+	log "github.com/sirupsen/logrus"
+
 	"github.com/ichiban/jesi/balance"
 	"github.com/ichiban/jesi/cache"
 	"github.com/ichiban/jesi/conditional"
+	"github.com/ichiban/jesi/control"
 	"github.com/ichiban/jesi/embed"
-	log "github.com/sirupsen/logrus"
 )
 
 var version string
 
 func main() {
-	var port int
+	var proxy Proxy
 	var backends balance.BackendPool
-	var max uint64
+	var store cache.Store
 	var verbose bool
+	var controls control.ClientPool
 
-	flag.IntVar(&port, "port", 8080, "port number")
+	flag.IntVar(&proxy.Port, "port", 8080, "port number")
 	flag.Var(&backends, "backend", "backend servers")
-	flag.Uint64Var(&max, "max", 64, "max cache size in MB")
+	flag.IntVar(&store.Max, "max", 64, "max cache size in MB")
 	flag.BoolVar(&verbose, "verbose", false, "log extra information")
+	flag.Var(&controls, "control", "control servers")
 	flag.Parse()
 
 	if verbose {
@@ -32,14 +36,16 @@ func main() {
 
 	go backends.Run(nil)
 
+	controls.Backends = &backends
+	controls.Store = &store
+	go controls.Run()
+
 	handler := http.Handler(&balance.Handler{
 		BackendPool: &backends,
 	})
 	handler = &cache.Handler{
-		Next: handler,
-		Store: cache.Store{
-			Max: max * 1024 * 1024,
-		},
+		Next:  handler,
+		Store: &store,
 	}
 	handler = &embed.Handler{
 		Next: handler,
@@ -50,15 +56,26 @@ func main() {
 
 	log.WithFields(log.Fields{
 		"version": version,
-		"port":    port,
-		"max":     max,
+		"port":    proxy.Port,
+		"max":     store.Max,
 		"verbose": verbose,
 	}).Info("Started a server")
 
+	proxy.Handler = handler
+	proxy.Run()
+}
+
+type Proxy struct {
+	Port    int
+	Handler http.Handler
+}
+
+func (p *Proxy) Run() {
 	server := http.Server{
-		Addr:    fmt.Sprintf(":%d", port),
-		Handler: handler,
+		Addr:    fmt.Sprintf(":%d", p.Port),
+		Handler: p.Handler,
 	}
+
 	if err := server.ListenAndServe(); err != nil {
 		log.WithFields(log.Fields{
 			"error": err,
